11/6/14 - 2
Initial upload of Terry's folder of the index and project and project2
pages. These pages allow the user to select the classes to generate the
schedule from.

Also uploading changes to daylogic to prevent system failure. Added
default return true values for when there is no conflict on the daysl