11/11/14
Changes made in the last 5 days. Added cdays to list the days for each
schedule. Fixed daylogic and added a missing logic piece. Fixed
get_times further to where it now generates the schedules correctly.
Added to get_times to now list the days correctly and to start counting
the number of hours the user is taking per semester.
Upload all terry's changes from the weekend. Note: He needs to change
his front end. Changes needed: Array needs to be of names not numbers,
array needs to be sent to Andrew's PHP page in a new window.

Additions to the system needed:
Page to display schedules.
Page to delete all schedules of a user - this page will be triggered by
javascript on the display schedule page

Database changes: No structural changes, but I had to redesign the
program to turn the data into sql queries because we forgot to add the
credit hours, which need to be in the database for the schedule to be
truly useful.

Jeff and Christine are working on the data to get it into form for my
program.