Final Prototype
Finished the prototype.

Changed the structure of the system to a more uniform structure.

All files are in the system directory now and then organized into the
proper subdirectories. All Terry's files have been moved to the main
directory to create a team appearance.

All mysql and mysqli connection strings have been replaced with an
include of the connection file in the includes directory. The connection
file therefore has been updated to work with the mysqli used in Terry's
code.

Added redirects to open pages on the fly. These are stored in the
includes section.

Added display page to the main directory. This page takes a student
number arguement and displays the possible schedules and then allows the
user to select one and see the details.

Updated links within Terry's code to reflect the changes in structure.
Mostly could not have been avoided once the structure changed, but one
or two could have been avoided by using relative links instead of hard
links.

Also updated the necessary links to work on localhost as well as the
connection string.

Working on localhost allowed for an easier debugging process due to
error messages. On web host error messages were not displayed.